Refactored solution
It doesn't make sense for the domain to know how to access files and
read the csvs. So now I removed the line and file reader related code
into a new project called Storage, whose sole purpose is to "access" the
persistence underneath - in our case the file system.
The domain, exposes an interface, repository that currently has only 2
read methods - settled and unsettled bets - . This is implemented in the
Storage by the Repository and the portal through ninject does the right
binding. So now in the domain we have only domain logic and in storage
only storage related operations.
Also updated the unit tests, moved the storage related ones to the
storage.